fix all issues based on the new compiler flags
- change timer variables from `uin64_t` to `int64_t`
- fix conversion issues form unsigned to signed
- remove usage of non C++ variable length arrays e.g `int x[N]` where `N` is a runtime value

+++ xmrstak/backend/amd/amd_gpu/gpu.hpp
@@ -46,7 +46,7 @@ uint32_t getNumPlatforms();
 int getAMDPlatformIdx();
 std::vector<GpuContext> getAMDDevices(int index);
 
-size_t InitOpenCL(GpuContext* ctx, size_t num_gpus, size_t platform_idx);
+int InitOpenCL(GpuContext* ctx, size_t num_gpus, size_t platform_idx);
 size_t XMRSetJob(GpuContext* ctx, uint8_t* input, size_t input_len, uint64_t target);
 size_t XMRRunJob(GpuContext* ctx, cl_uint* HashOutput);
