+++ xmrstak/backend/amd/jconf.hpp
@@ -25,7 +25,7 @@ public:
 		size_t index;
 		size_t intensity;
 		size_t w_size;
-		long long cpu_aff;
+		int32_t cpu_aff;
 		bool stridedIndex;
 	};
